Change float to int

+++ src/entities/Entity.java
@@ -4,7 +4,7 @@ import java.awt.*;
 import java.awt.image.BufferedImage;
 
 public abstract class Entity {
-    protected float worldX, worldY, speed;
+    protected int worldX, worldY, speed;
     protected int width, height;
     protected BufferedImage up1, up2, left1, left2, down1, down2, right1, right2;
     // TODO: once have pathfinding may cause problem because of String not Integer
@@ -16,7 +16,7 @@ public abstract class Entity {
     protected int hitboxDefaultX, hitboxDefaultY;
     protected boolean collision;
 
-    public Entity(float worldX, float worldY, float speed, int width, int height) {
+    public Entity(int worldX, int worldY, int speed, int width, int height) {
         this.worldX = worldX;
         this.worldY = worldY;
         this.speed = speed;
@@ -28,15 +28,15 @@ public abstract class Entity {
         return direction;
     }
 
-    public float getWorldX() {
+    public int getWorldX() {
         return worldX;
     }
 
-    public float getWorldY() {
+    public int getWorldY() {
         return worldY;
     }
 
-    public float getSpeed() {
+    public int getSpeed() {
         return speed;
     }
